backend/requirements.txt

fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
httpx==0.28.1
python-multipart==0.0.20
motor==3.6.1

# Medical file handling (MVP best-effort)
pydicom==2.4.4
nibabel==5.3.2
pillow==11.0.0
opencv-python-headless==4.10.0.84

# Providers
kaggle==1.6.14

# Profiling & ML
torch==2.5.1
torchvision==0.20.1
hdbscan==0.8.40
scikit-learn==1.6.0
timm==1.0.12


